refactor(test): split the oversized permalink/UPI spec files by concern
Pure test-file reorganization — no production code change. The four
>1000-line permalink/UPI spec files are split by concern into sibling
files, each under the 1000-line boundary, with every test case preserved
verbatim (byte-for-byte moves; it()-title sets unchanged).

- permalink.controller.spec.ts (2008) -> resolve/read (983) + .list (414) + .write (766)
- permalink.application.service.spec.ts (1598) -> default/resolve (704) + .create-delete (562) + .list (394)
- upi-collection.service.spec.ts (1110) -> write (606) + .list (577)
- permalink.repository.spec.ts (1076) -> crud/finders (728) + .pagination (410)

10 suites / 223 tests green; oxlint clean. Closes #641.
